Move IQuickActivate to Ole32
