+++ src/System.Windows.Forms.Design/src/System.Windows.Forms.Design.csproj
@@ -143,10 +143,6 @@
     <Compile Include="..\..\Common\src\Interop\Ole32\Interop.OLEWHICHMK.cs" Link="Interop\Ole32\Interop.OLEWHICHMK.cs" />
     <Compile Include="..\..\Common\src\Interop\Ole32\Interop.PARAMDESC.cs" Link="Interop\Ole32\Interop.PARAMDESC.cs" />
     <Compile Include="..\..\Common\src\Interop\Ole32\Interop.PARAMFLAG.cs" Link="Interop\Ole32\Interop.PARAMFLAG.cs" />
-    <Compile Include="..\..\Common\src\Interop\Ole32\Interop.POINTERINACTIVE.cs" Link="Interop\Ole32\Interop.POINTERINACTIVE.cs" />
-    <Compile Include="..\..\Common\src\Interop\Ole32\Interop.QACONTAINER.cs" Link="Interop\Ole32\Interop.QACONTAINER.cs" />
-    <Compile Include="..\..\Common\src\Interop\Ole32\Interop.QACONTAINERFLAGS.cs" Link="Interop\Ole32\Interop.QACONTAINERFLAGS.cs" />
-    <Compile Include="..\..\Common\src\Interop\Ole32\Interop.QACONTROL.cs" Link="Interop\Ole32\Interop.QACONTROL.cs" />
     <Compile Include="..\..\Common\src\Interop\Ole32\Interop.RevokeDragDrop.cs" Link="Interop\Ole32\Interop.RevokeDragDrop.cs" />
     <Compile Include="..\..\Common\src\Interop\Ole32\Interop.STATFLAG.cs" Link="Interop\Ole32\Interop.STATFLAG.cs" />
     <Compile Include="..\..\Common\src\Interop\Ole32\Interop.STATSTG.cs" Link="Interop\Ole32\Interop.STATSTG.cs" />
